feat: healthprobe wiring, context-aware wallet client, lint fixes
- cmd/blrcs-mcpd: replace stub /healthz + /readyz handlers with
  structured healthprobe.Probe; adds ledger + storage readiness checks
  returning JSON reports with per-check status and duration
- openid4vci: add FetchCredentialCtx/FetchMetadataCtx/FetchJWKSCtx
  with context.Context for cancellation and timeout; original methods
  delegate to these via context.Background() (backward-compatible)
- .goreleaser.yaml: add blrcs-mcp stdio binary to release builds
- scitt/consistency.go: eliminate int(uint64) round-trip in subProof
  using math/bits.Len64 for direct power-of-2 computation
- lint: remove unused versionTimeValid (didwebvh), unnecessary type
  casts (cose_receipt, vctmeta), De Morgan simplification (httpchain),
  unused time import; all verified clean with golangci-lint v2

+++ blrcs/scitt/cose_receipt.go
@@ -59,8 +59,8 @@ func IssueCOSEReceipt(r *Receipt, tsPriv ed25519.PrivateKey, tsID string) ([]byt
 	}
 
 	payload, err := cbor.Marshal(map[int]any{
-		cbrLeafIndex:  uint64(r.LeafIndex),
-		cbrTreeSize:   uint64(r.TreeSize),
+		cbrLeafIndex:  r.LeafIndex,
+		cbrTreeSize:   r.TreeSize,
 		cbrRootHash:   rootBytes,
 		cbrAuditPath:  pathAny,
 		cbrTSID:       tsID,
